Refactor functional tests
